apps/api/requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
pydantic>=2.7
SQLAlchemy>=2.0
httpx>=0.27
requests>=2.31
python-dotenv>=1.0
websockets>=12.0
claude-agent-sdk==0.1.4
docker>=7.0
cryptography>=42.0
openai>=1.40
unidiff>=0.7
aiohttp>=3.9
rich>=13.0
python-multipart>=0.0.6
sqlglot>=25.0.0
pyOpenSSL>=24.0
pyyaml>=6.0
Jinja2>=3.1.6
openpyxl>=3.1.0
pandas==2.3.3
scikit-learn==1.6.1
backtrader==1.9.78.123
